@@ -1,87 +1,85 @@
 {
-    "name": "admin-portal",
-    "version": "1.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "fast_start": "NG_PERSISTENT_BUILD_CACHE=1 ng serve --port=5600",
-        "build": "ng build",
-        "prod": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration production",
-        "staging": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration staging",
-        "qa": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration qa",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular-material-extensions/password-strength": "^8.2.1",
-        "@angular/animations": "~12.1.1",
-        "@angular/cdk": "^12.2.8",
-        "@angular/common": "~12.1.1",
-        "@angular/compiler": "~12.1.1",
-        "@angular/core": "~12.1.1",
-        "@angular/forms": "~12.1.1",
-        "@angular/material": "^12.2.8",
-        "@angular/platform-browser": "~12.1.1",
-        "@angular/platform-browser-dynamic": "~12.1.1",
-        "@angular/router": "~12.1.1",
-        "@types/crypto-js": "^4.0.1",
-        "@types/moment-timezone": "^0.5.30",
-        "@types/socket.io-client": "^1.4.36",
-        "angular-archwizard": "^7.0.0",
-        "angular-datatables": "^12.0.0",
-        "b64-to-blob": "^1.2.19",
-        "bootstrap": "^4.6.0",
-        "crypto-js": "^4.0.0",
-        "datatables.net": "^1.10.25",
-        "datatables.net-colreorder": "^1.5.4",
-        "datatables.net-colreorder-dt": "^1.5.4",
-        "datatables.net-dt": "^1.1.0",
-        "datatables.net-select": "^1.3.3",
-        "datatables.net-select-dt": "^1.3.3",
-        "highcharts": "^9.3.0",
-        "highcharts-angular": "^3.0.0",
-        "jquery": "^3.6.0",
-        "moment-timezone": "^0.5.33",
-        "ng-block-ui": "^3.0.2",
-        "ng-pick-datetime": "^7.0.0",
-        "ngx-bootstrap": "^6.2.0",
-        "ngx-daterangepicker-material": "^2.4.2",
-        "ngx-dropzone": "^3.0.0",
-        "ngx-editor": "^10.1.0",
-        "ngx-image-cropper": "^4.0.0",
-        "ngx-lightbox": "^2.4.1",
-        "ngx-mask": "^12.0.0",
-        "ngx-mat-select-search": "^3.3.3",
-        "ngx-pagination": "^5.1.1",
-        "ngx-toastr": "^14.0.0",
-        "rxjs": "^6.5.3",
-        "rxjs-compat": "^6.6.7",
-        "socket.io-client": "^2.4.0",
-        "tslib": "^2.3.0",
-        "zone.js": "~0.11.4"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~12.1.1",
-        "@angular/cli": "~12.1.1",
-        "@angular/compiler-cli": "~12.1.1",
-        "@types/datatables.net": "^1.10.20",
-        "@types/datatables.net-select": "^1.2.7",
-        "@types/jasmine": "~3.8.1",
-        "@types/jquery": "^3.5.6",
-        "@types/node": "^16.3.1",
-        "codelyzer": "^6.0.2",
-        "jasmine-core": "~3.8.0",
-        "jasmine-spec-reporter": "~7.0.0",
-        "karma": "~6.3.4",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.0.3",
-        "karma-jasmine": "~4.0.1",
-        "karma-jasmine-html-reporter": "^1.7.0",
-        "protractor": "~7.0.0",
-        "ts-node": "~10.1.0",
-        "tslint": "~6.1.3",
-        "typescript": "~4.3.5"
-    }
+  "name": "admin-portal",
+  "version": "1.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "fast_start": "NG_PERSISTENT_BUILD_CACHE=1 ng serve --port=4200",
+    "build": "ng build",
+    "prod": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration production",
+    "staging": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration staging",
+    "qa": "node --max_old_space_size=4096 node_modules/@angular/cli/bin/ng build --configuration qa",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@types/crypto-js": "^4.0.1",
+    "@types/moment-timezone": "^0.5.30",
+    "@types/socket.io-client": "^1.4.36",
+    "angular-datatables": "^14.0.1",
+    "b64-to-blob": "^1.2.19",
+    "bootstrap": "^4.6.0",
+    "crypto-js": "^4.0.0",
+    "datatables.net": "^1.13.3",
+    "datatables.net-colreorder": "^1.6.1",
+    "datatables.net-colreorder-dt": "^1.6.1",
+    "datatables.net-dt": "^1.13.3",
+    "datatables.net-select": "^1.6.1",
+    "datatables.net-select-dt": "^1.6.1",
+    "highcharts": "^9.3.0",
+    "highcharts-angular": "^3.0.0",
+    "jquery": "^3.6.0",
+    "moment-timezone": "^0.5.33",
+    "ng-block-ui": "^3.0.2",
+    "ng-pick-datetime": "^7.0.0",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-daterangepicker-material": "^2.4.2",
+    "ngx-dropzone": "^3.0.0",
+    "ngx-editor": "^15.3.0",
+    "ngx-image-cropper": "^4.0.0",
+    "ngx-lightbox": "^2.4.1",
+    "ngx-mask": "^12.0.0",
+    "ngx-mat-select-search": "^3.3.3",
+    "ngx-pagination": "^5.1.1",
+    "ngx-toastr": "^14.0.0",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.6.7",
+    "socket.io-client": "^2.4.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@types/datatables.net": "^1.10.20",
+    "@types/datatables.net-select": "^1.2.7",
+    "@types/jasmine": "~3.8.1",
+    "@types/jquery": "^3.5.6",
+    "@types/node": "^16.3.1",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.8.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.4",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.1.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.6.4"
+  }
 }
