@@ -1,5 +1,5 @@
 import { EventEmitter, Component, OnInit, Output, ChangeDetectorRef } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder, FormArray } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder, FormArray } from '@angular/forms';
 import { Helper } from 'src/app/services/helper.service';
 import { BsModalRef } from 'ngx-bootstrap/modal'
 import { AttributesAddEditModalService } from './attributes-add-edit-modal.service';
@@ -14,7 +14,7 @@ import { HttpClient } from '@angular/common/http';
 export class AttributesAddEditModalComponent implements OnInit {
   @Output() onEventCompleted: EventEmitter<any> = new EventEmitter();
   modalEvent: any;
-  attributesForm: FormGroup;
+  attributesForm: UntypedFormGroup;
   imageUrl: any = '../../../../../assets/img/no_preview.png';
   selectedImageFile: any
   attributeDetails:any;
@@ -24,20 +24,30 @@ export class AttributesAddEditModalComponent implements OnInit {
     private _http: HttpClient,
     private _helper:Helper,
     private _bsModalRef:BsModalRef,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private _changeDetectorRef:ChangeDetectorRef,
     private _tcAddEditModalService: AttributesAddEditModalService){
 
     this.attributesForm = this.formBuilder.group({
-      'id':new FormControl(null, []),
-      'name': new FormControl(null, [Validators.required]),
-      'is_active': new FormControl(null, [])
+      'id':new UntypedFormControl(null, []),
+      'name': new UntypedFormControl(null, [Validators.required]),
+      'is_active': new UntypedFormControl(null, []),
+      'name_fr': new UntypedFormControl(null,[]),
+      'name_nl': new UntypedFormControl(null,[]),
+      'name_es': new UntypedFormControl(null,[]),
+      'name_pt': new UntypedFormControl(null,[]),
+      'element_type': new UntypedFormControl('dropdown',[Validators.required])
     });
   }
 
   get id() { return this.attributesForm.get('id'); }
   get name() { return this.attributesForm.get('name'); }
   get is_active() { return this.attributesForm.get('is_active'); }
+  get name_fr() { return this.attributesForm.get('name_fr'); }
+  get name_nl() { return this.attributesForm.get('name_nl'); }
+  get name_es() { return this.attributesForm.get('name_es'); }
+  get name_pt() { return this.attributesForm.get('name_pt'); }
+  get element_type() { return this.attributesForm.get('element_type'); }
 
   ngOnInit(): void {
     let details = this._tcAddEditModalService.getData();
@@ -61,7 +71,12 @@ export class AttributesAddEditModalComponent implements OnInit {
     this.attributesForm.patchValue({
       id:this.attributeDetails.id,
       name:this.attributeDetails.name,
-      is_active:this.attributeDetails.is_active
+      element_type:this.attributeDetails.element_type,
+      is_active:this.attributeDetails.is_active,
+      name_fr:this.attributeDetails.name_fr,
+      name_nl:this.attributeDetails.name_nl,
+      name_es:this.attributeDetails.name_es,
+      name_pt:this.attributeDetails.name_pt
     });
   }
  
