@@ -1,5 +1,5 @@
 import { EventEmitter, Component, OnInit, Output, ChangeDetectorRef } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder, FormArray } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder, UntypedFormArray } from '@angular/forms';
 import { Helper } from 'src/app/services/helper.service';
 import { BsModalRef } from 'ngx-bootstrap/modal'
 import { SubcategoriesAddEditModalService } from './subcategories-add-edit-modal.service';
@@ -14,7 +14,7 @@ import { HttpClient } from '@angular/common/http';
 export class SubcategoriesAddEditModalComponent implements OnInit {
   @Output() onEventCompleted: EventEmitter<any> = new EventEmitter();
   modalEvent: any;
-  SubcategoryForm: FormGroup;
+  SubcategoryForm: UntypedFormGroup;
 
   imageUrl: any = '../../../../../assets/img/no_preview.png';
   selectedImageFile: any
@@ -25,18 +25,28 @@ export class SubcategoriesAddEditModalComponent implements OnInit {
     private _http: HttpClient,
     private _helper:Helper,
     private _bsModalRef:BsModalRef,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private _changeDetectorRef:ChangeDetectorRef,
     private _tcAddEditModalService: SubcategoriesAddEditModalService
   ) {
     this.SubcategoryForm = this.formBuilder.group({
-      'category_id':new FormControl(null, []),
-      'id':new FormControl(null, []),
-      'name': new FormControl(null, [Validators.required]),
-      'is_active': new FormControl(null, []),
-      'description': new FormControl(null, []),
-      'image_url': new FormControl(null, []),
-      'attributes':new FormArray([],[])
+      'category_id':new UntypedFormControl(null, []),
+      'id':new UntypedFormControl(null, []),
+      'name': new UntypedFormControl(null, [Validators.required]),
+      'is_active': new UntypedFormControl(null, []),
+      'description': new UntypedFormControl(null, []),
+      'image_url': new UntypedFormControl(null, []),
+      'min_shipping_weight': new UntypedFormControl(null, []),
+      'max_shipping_weight': new UntypedFormControl(null, []),
+      'attributes':new UntypedFormArray([]),
+      'name_fr': new UntypedFormControl(null,[]),
+      'name_nl': new UntypedFormControl(null,[]),
+      'name_es': new UntypedFormControl(null,[]),
+      'name_pt': new UntypedFormControl(null,[]),
+      'description_fr': new UntypedFormControl(null,[]),
+      'description_nl': new UntypedFormControl(null,[]),
+      'description_es': new UntypedFormControl(null,[]),
+      'description_pt': new UntypedFormControl(null,[]),
     });
   }
 
@@ -46,6 +56,17 @@ export class SubcategoriesAddEditModalComponent implements OnInit {
   get is_active() { return this.SubcategoryForm.get('is_active'); }
   get description() { return this.SubcategoryForm.get('description'); }
   get image_url() { return this.SubcategoryForm.get('image_url'); }
+  get min_shipping_weight() { return this.SubcategoryForm.get('min_shipping_weight'); }
+  get max_shipping_weight() { return this.SubcategoryForm.get('max_shipping_weight'); }
+  get name_fr() { return this.SubcategoryForm.get('name_fr'); }
+  get name_nl() { return this.SubcategoryForm.get('name_nl'); }
+  get name_es() { return this.SubcategoryForm.get('name_es'); }
+  get name_pt() { return this.SubcategoryForm.get('name_pt'); }
+
+  get description_fr() { return this.SubcategoryForm.get('description_fr'); }
+  get description_nl() { return this.SubcategoryForm.get('description_nl'); }
+  get description_es() { return this.SubcategoryForm.get('description_es'); }
+  get description_pt() { return this.SubcategoryForm.get('description_pt'); }
 
   ngOnInit(): void {
     this.getAllAttributesData();
@@ -120,26 +141,78 @@ export class SubcategoriesAddEditModalComponent implements OnInit {
       category_id:this.subcategoryDetails.category_id,
       name:this.subcategoryDetails.name,
       is_active:this.subcategoryDetails.is_active,
-      description:this.subcategoryDetails.description
+      description:this.subcategoryDetails.description,
+      min_shipping_weight:this.subcategoryDetails.min_shipping_weight,
+      max_shipping_weight:this.subcategoryDetails.max_shipping_weight,
+      name_fr:this.subcategoryDetails.name_fr,
+      name_nl:this.subcategoryDetails.name_nl,
+      name_es:this.subcategoryDetails.name_es,
+      name_pt:this.subcategoryDetails.name_pt,
+      description_fr:this.subcategoryDetails.description_fr,
+      description_nl:this.subcategoryDetails.description_nl,
+      description_es:this.subcategoryDetails.description_es,
+      description_pt:this.subcategoryDetails.description_pt
     });
     this.imageUrl = this.subcategoryDetails.image ? environment.api_url + this.subcategoryDetails.image : `../../../../../assets/img/no_preview.png`;
-    const attributesControl = this.SubcategoryForm.get('attributes') as FormArray;
+    const attributesControl = this.SubcategoryForm.get('attributes') as UntypedFormArray;
     if(this.subcategoryDetails.attributes){
-      this.subcategoryDetails.attributes.forEach((item:any)=>{
-        let attributeFormGroup = new FormGroup({
-          'attribute_id':new FormControl(item.attribute_id, [Validators.required]),
-          'values':new FormControl(item.values_ids, [Validators.required]),
+
+      this.subcategoryDetails.attributes.forEach((item:any,index:number)=>{
+        let valArray:any  = [];
+        item.values.forEach((val:any,valIndex:number)=>{
+          let chArray :any = [];
+          if(val.children && val.children.length>0){
+             val.children.forEach((child:any,childIndex:number)=>{
+                  let childValArray= child.values.slice().map((ch:any)=> ch.attribute_value_id);
+                  let childFormGroup =  new UntypedFormGroup({
+                    'child_attribute_id': new UntypedFormControl(child.attribute_id, []),
+                    'child_attribute_values': new UntypedFormControl(childValArray, [])
+                  });
+                  chArray.push(childFormGroup);
+             });
+          }
+          let valFormGroup= new UntypedFormGroup({
+            'value_id': new UntypedFormControl(val.attribute_value_id, [Validators.required]),
+            'children':new UntypedFormArray(chArray)
+          });
+          valArray.push(valFormGroup);
         });
+        let attributeFormGroup = new UntypedFormGroup({
+          'attribute_id':new UntypedFormControl(item.attribute_id, [Validators.required]),
+          'values':new UntypedFormArray(valArray)
+        });
+        console.log('attributeFormGroup=',attributeFormGroup);
         attributesControl.push(attributeFormGroup);
       });
     }
   }
 
+  childAttrChange(attrIndex:number,valueIndex:number,childIndex:number,event:any){
+    this.childrenArray(attrIndex,valueIndex).at(childIndex).patchValue({
+      child_attribute_id:event.value
+    });
+    let valuesss= this.getValuesFromAttribute('value',event.value);
+  }
+
+  childAttrValueChange(attrIndex:number,valueIndex:number,childIndex:number,event:any){
+    this.childrenArray(attrIndex,valueIndex).at(childIndex).patchValue({
+      child_attribute_values:event.value
+    });
+  }
+
+
   attrChange(index:number,event:any){
-      this.attributes().at(index).patchValue({
-        attribute_id:event.value
-      });
-      let valuesss= this.getValuesFromAttribute('value',event.value);
+    this.attributes().at(index).patchValue({
+      attribute_id:event.value
+    });
+    let valuesss= this.getValuesFromAttribute('value',event.value);
+    this.clearFormArray(this.valuesInputArray(index));
+  }
+
+  clearFormArray = (formArray: UntypedFormArray) => {
+    while (formArray.length !== 0) {
+      formArray.removeAt(0)
+    }
   }
 
   getValuesFromAttribute(key:string,value:number){
@@ -158,27 +231,86 @@ export class SubcategoriesAddEditModalComponent implements OnInit {
       return values;
   }
 
-  /*-----------------------------ATTRIBUTES --------------------------------*/
-  addAttributeInput(){
-    this.attributes().push(this.newAttributeInput());
-    this.SubcategoryForm.updateValueAndValidity();
-  }
+    /*-----------------------------ATTRIBUTES --------------------------------*/
+    addAttributeInput(){
+      this.attributes().push(this.newAttributeInput());
+      this.SubcategoryForm.updateValueAndValidity();
+    }
 
-  newAttributeInput(): FormGroup{
-    return new FormGroup({
-      'attribute_id': new FormControl('', [Validators.required]),
-      'values': new FormControl([], [Validators.required]),
-    });
-  }
+    newAttributeInput(): UntypedFormGroup{
+      return new UntypedFormGroup({
+        'attribute_id': new UntypedFormControl('', [Validators.required]),
+        'values': new UntypedFormArray([], [Validators.required]),
+      });
+    }
 
-  removeAttributeInput(empIndex:number){
-    this.attributes().removeAt(empIndex);
-  }
+    removeAttributeInput(attrIndex:number){
+      this.attributes().removeAt(attrIndex);
+    }
 
-  attributes(): FormArray {
-      return this.SubcategoryForm.get("attributes") as FormArray
-  }
-  /*-----------------------------END OF ATTRIBUTES --------------------------------*/
+    attributes(): UntypedFormArray {
+        return this.SubcategoryForm.get("attributes") as UntypedFormArray
+    }
+    /*-----------------------------END OF ATTRIBUTES ---------------------*/
+
+    /*-----------------------------VALUES --------------------------------*/
+    addValueInput(attrIndex:number,attrValue:number){
+      this.valuesInputArray(attrIndex).push(this.newValueInput(attrValue));
+      this.SubcategoryForm.updateValueAndValidity();
+    }
+
+    newValueInput(attrValue:number): UntypedFormGroup{
+      return new UntypedFormGroup({
+        'value_id': new UntypedFormControl(attrValue, [Validators.required]),
+        'children':new UntypedFormArray([], [])
+      });
+    }
+
+    removeValueInput(attrIndex:number,valueIndex:number){
+      this.valuesInputArray(attrIndex).removeAt(valueIndex);
+    }
+
+    valuesInputArray(attrIndex:number): UntypedFormArray {
+      return this.attributes().at(attrIndex).get('values') as UntypedFormArray
+    }
+
+    getcheckBoxValue(attrIndex:number,inputVal:number) {
+       let checked = this.valuesInputArray(attrIndex) ? (this.valuesInputArray(attrIndex)?.value.filter((itm:any)=>itm.value_id == inputVal).length>0?true:false):false;
+       return checked;
+    }
+
+    /*-----------------------------END OF CHILD ATTRIBUTES --------------------------------*/
+
+    /*--------------------------------- CHILDREN ----------------------------------*/
+    newChildInput(){
+      return new UntypedFormGroup({
+        'child_attribute_id': new UntypedFormControl('', []),
+        'child_attribute_values': new UntypedFormControl([], [])
+      });
+    }
+
+    addNewChildToAttrChildren(attrIndex:number,valueIndex:number){
+      this.childrenArray(attrIndex,valueIndex).push(this.newChildInput());
+      this.SubcategoryForm.updateValueAndValidity();
+    }
+
+    removeChildFromAttrChildren(attrIndex:number,valueIndex:number,childIndex:number){
+      this.childrenArray(attrIndex,valueIndex).removeAt(childIndex);
+    }
+
+    childrenArray(attrIndex:number,valueIndex:number){
+      return this.valuesInputArray(attrIndex).at(valueIndex).get('children') as UntypedFormArray
+    }
+    /*-------------------------------- END CHILDREN -------------------------------*/
+
+
+    setAttributeValue(attrIndex:number,valueIndex:number,attrValue:number,checked:boolean){
+      if(checked){
+          this.addValueInput(attrIndex,attrValue);
+      }else{
+          this.removeValueInput(attrIndex,valueIndex);
+      }
+    }
 
   closeModal(){
     this._bsModalRef.hide();
