@@ -1,5 +1,5 @@
-import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { Component, OnInit, ChangeDetectorRef, OnDestroy } from '@angular/core';
+import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 import { Toastr } from 'src/app/services/toastr.service';
 import { Router } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
@@ -11,10 +11,20 @@ import { Editor } from 'ngx-editor';
   templateUrl: './create-faq.component.html',
   styleUrls: ['./create-faq.component.scss']
 })
-export class CreateFaqComponent implements OnInit {
+export class CreateFaqComponent implements OnInit,OnDestroy {
   editor: Editor;
-  answerText:string='';
-  public FAQForm!: FormGroup;
+  editor_fr: Editor;
+  editor_nl: Editor;
+  editor_es: Editor;
+  editor_pt: Editor;
+  
+  answerText:string=''; 
+  answerText_fr:string='';
+  answerText_nl:string='';
+  answerText_es:string='';
+  answerText_pt:string='';
+  
+  public FAQForm!: UntypedFormGroup;
   public FAQgroup:any=[];
   public eventInfo!: string;
 
@@ -25,6 +35,11 @@ export class CreateFaqComponent implements OnInit {
     private _http: HttpClient,
     private _helper: Helper){
     this.editor = new Editor();
+    this.editor_fr = new Editor();
+    this.editor_nl = new Editor();
+    this.editor_es = new Editor();
+    this.editor_pt = new Editor();
+    
   }
 
   getFAQGroups(){
@@ -38,26 +53,55 @@ export class CreateFaqComponent implements OnInit {
 
   ngOnInit() {
     this.getFAQGroups();
-    this.FAQForm = new FormGroup({
-      'group_id': new FormControl([],[Validators.required]),
-      'question': new FormControl(null, [Validators.required]),
-      'answer': new FormControl(null, [Validators.required]),
-      'is_active': new FormControl(true)
+    this.FAQForm = new UntypedFormGroup({
+      'group_id': new UntypedFormControl([],[Validators.required]),
+      'question': new UntypedFormControl(null, [Validators.required]),
+      'answer': new UntypedFormControl(null, [Validators.required]),
+      'question_fr': new UntypedFormControl(null,[]),
+      'answer_fr': new UntypedFormControl(null,[]),
+      'question_nl': new UntypedFormControl(null,[]),
+      'answer_nl': new UntypedFormControl(null,[]),
+      'question_es': new UntypedFormControl(null,[]),
+      'answer_es': new UntypedFormControl(null,[]),
+      'question_pt': new UntypedFormControl(null,[]),
+      'answer_pt': new UntypedFormControl(null,[]),
+      'is_active': new UntypedFormControl(true)
     });
   }
 
   ngOnDestroy(): void {
     this.editor.destroy();
+    this.editor_fr.destroy();
+    this.editor_nl.destroy();
+    this.editor_es.destroy();
+    this.editor_pt.destroy();
   }
 
   get group_id() { return this.FAQForm.get('group_id'); }
   get question() { return this.FAQForm.get('question'); }
   get answer() { return this.FAQForm.get('answer'); }
+
+  get question_fr() { return this.FAQForm.get('question_fr'); }
+  get answer_fr() { return this.FAQForm.get('answer_fr'); }
+  
+  get question_nl() { return this.FAQForm.get('question_nl'); }
+  get answer_nl() { return this.FAQForm.get('answer_nl'); }
+  
+  get question_es() { return this.FAQForm.get('question_es'); }
+  get answer_es() { return this.FAQForm.get('answer_es'); }
+  
+  get question_pt() { return this.FAQForm.get('question_pt'); }
+  get answer_pt() { return this.FAQForm.get('answer_pt'); }
+  
   get is_active() { return this.FAQForm.get('is_active'); }
 
   patchAnswer(){
     this.FAQForm.patchValue({
-      answer:this.answerText
+      answer:this.answerText,
+      answer_fr:this.answerText_fr,
+      answer_nl:this.answerText_nl,
+      answer_es:this.answerText_es,
+      answer_pt:this.answerText_pt
     });
   }
 
