@@ -69,7 +69,6 @@ export class NavbarComponent implements OnInit, OnDestroy {
 			{ route: '/admin/products', iconName: 'fa-golf-ball', displayName: 'Products' },
 			{ route: '/admin/top-listing-subscriptions', iconName: 'fa-golf-ball', displayName: 'Top Listing Subscriptions' },
 			{ route: '/admin/dealer-subscriptions',  iconName: 'fa-users', displayName: 'Dealers Account Subscriptions'},
-			{ route: '/admin/bannersets' , iconName: 'fa-images', displayName: 'BannerSets' },
 			{ route: '/admin/categories', iconName: 'fa-bars', displayName: 'Categories' },
 			{ route: '/admin/plans', iconName: 'fa-list-alt', displayName: 'Subscription Plans' },
 			{ route: '/admin/attributes', iconName: 'fa-bars', displayName: 'Attributes' },
@@ -84,12 +83,13 @@ export class NavbarComponent implements OnInit, OnDestroy {
 					{ route: '/admin/faq', iconName: 'fa-question-circle', displayName: 'FAQ' },
 					{ route: '/admin/faq-group', iconName: 'fa-question-circle', displayName: 'FAQ Group' },
 					{ route: '/admin/settings', iconName: 'fa-cog', displayName: 'Settings' },
+					/*{ route: '/admin/shipping_pricing',iconName:'fa-cog',displayName:'Shipping Pricing'},*/
+					{ route: '/admin/shipping_insurance', iconName: 'fa-cog', displayName: 'Shipping insurance' },
 					{ route: '/admin/article-categories', iconName: 'fa-bars', displayName: 'Article Categories' },
 					{ route: '/admin/articles', iconName: 'fa-newspaper', displayName: 'Articles' },
 				]
 			},
-			{ route: '/admin/reports', iconName: 'fa-file', displayName: 'Reports' }
-		]
+			/*{ route: '/admin/reports', iconName: 'fa-file', displayName: 'Reports' }*/		]
 	}
 
 	ngOnInit() {
