@@ -110,7 +110,7 @@ export class AttributeValuesComponent implements OnInit,AfterViewInit,OnDestroy 
               let url = environment.api_url + data;
               return `<img src=${url} height="80" width="80" />`;
             } else {
-              return ``;
+              return `-`;
             }
           }
         },
@@ -118,6 +118,30 @@ export class AttributeValuesComponent implements OnInit,AfterViewInit,OnDestroy 
           data: 'name',
           title: 'Value Name',
           className: 'text-center  font-weight-normal'
+        },
+        {
+          data: 'created_at',
+          title: 'Created At',
+          className: 'text-center  font-weight-normal',
+          render: (data) => {
+            if (data) {
+              return this._helper.getFormattedDate(data, 'DD/MM/YYYY');
+            } else {
+              return '<span>-</span>';
+            }
+          }
+        },
+        {
+          data: 'updated_at',
+          title: 'Updated At',
+          className: 'text-center  font-weight-normal',
+          render: (data) => {
+            if (data) {
+              return this._helper.getFormattedDate(data, 'DD/MM/YYYY');
+            } else {
+              return '<span>-</span>';
+            }
+          }
         },
         {
           data: 'is_active',
