@@ -113,7 +113,7 @@ export class SubscriptionPaymentHistoryComponent implements OnInit, AfterViewIni
             if (data) {
               return this._helper.getFormattedDate(data, 'DD/MM/YYYY');
             } else {
-              return '<span></span>';
+              return '<span>-</span>';
             }
           }
         },
@@ -122,7 +122,7 @@ export class SubscriptionPaymentHistoryComponent implements OnInit, AfterViewIni
           className: 'text-center',
           render: function (data: any, type: any, full: any) {
             return `
-            <button class="btn btn-sm btn-default m-0" receiptId=${full.receipt_url}>View Receipt</button>
+            <button class="btn btn-sm btn-primary m-0" receiptId=${full.receipt_url}>View Receipt</button>
             `;
           },
           orderable: false
