@@ -5,6 +5,7 @@ import { ProductInfoComponent } from './product-info.component';
 import { ChangeSequenceModule } from 'src/app/shared/change-sequence/change-sequence.module';
 import { Helper } from 'src/app/services/helper.service';
 import { SharedModule } from 'src/app/shared/shared.module';
+import { ReasonModalModule } from 'src/app/views/admin/common-components/reason-modal/reason-modal.module';
 
 
 @NgModule({
@@ -14,7 +15,8 @@ import { SharedModule } from 'src/app/shared/shared.module';
   imports: [
     SharedModule,
     ProductInfoRoutingModule,
-    ChangeSequenceModule
+    ChangeSequenceModule,
+    ReasonModalModule
   ],
   providers:[Helper]
 })
