@@ -4,9 +4,9 @@ export const environment = {
 	production: true,
 	current_environment: 'production',
 	encryption_key: 'YDF7IirWSIUxR4ijMHXQrpVQ1ZyI0bR6',
-	api_url: 'https://golfed-api.archesoftronic.com/',
-	client_app_url: 'http://localhost:4200/',
-	stripeKey: 'pk_test_51M9tzXE4Jg2znPuBAdh0ElxhTEwhLsk8NVEuaCzCWzR9qZ0LpaZWGXZiKtlsFb9KK6Iqw913Gr0eGLoTIcp0lYyK00k6fyw9kh',
-	websocket_url: 'https://golfed-api.archesoftronic.com/',
+	api_url: 'https://api.golfed-europe.eu/',
+	client_app_url: 'https://golfed-europe.eu/',
+	stripeKey: 'pk_live_51M9tzXE4Jg2znPuBV91ORKRghy9Qasmtk4OsA1y9yeQDwWSy1rFp5pBXBoYDgygHQIDUnhYjncKKQiUsLiiOyXMi00UYAfYvk6',
+	websocket_url: 'https://api.golfed-europe.eu/',
 	google_map_api_key: ''
 };
