@@ -1,12 +1,12 @@
 
-// stage
+// qa
 export const environment = {
 	production: false,
 	current_environment: 'qa',
 	encryption_key: 'YDF7IirWSIUxR4ijMHXQrpVQ1ZyI0bR6',
-	api_url: 'http://localhost:3001/',
-	client_app_url: 'http://localhost:4200/',
-	stripeKey: '',
-	websocket_url: 'http://localhost:3001/',
+	api_url: 'https://api-qa.golfed-europe.eu/',
+	client_app_url: 'https://golfed-qa.archesoftronix.in/',
+	stripeKey: 'pk_test_51M9tzXE4Jg2znPuBAdh0ElxhTEwhLsk8NVEuaCzCWzR9qZ0LpaZWGXZiKtlsFb9KK6Iqw913Gr0eGLoTIcp0lYyK00k6fyw9kh',
+	websocket_url: 'https://api-qa.golfed-europe.eu/',
 	google_map_api_key: ''
 };
