@@ -6,10 +6,10 @@ export const environment = {
 	production: false,
 	current_environment:'local',
   	encryption_key:'YDF7IirWSIUxR4ijMHXQrpVQ1ZyI0bR6',
-	api_url: 'http://localhost:3001/',
+	api_url: 'http://localhost:3002/',
   	client_app_url: 'http://localhost:4200/',
 	stripeKey: '',
-	websocket_url: 'http://localhost:3001/',
+	websocket_url: 'http://localhost:3002/',
   	google_map_api_key:''
 };
 
